+++ apps/docs/src/app/core/core-documentation.module.ts
@@ -400,8 +400,11 @@ import { SideNavigationObjectExampleComponent } from './component-docs/side-navi
 import { SideNavigationCondensedExampleComponent } from './component-docs/side-navigation/examples/side-navigation-condensed-example/side-navigation-condensed-example.component';
 import { SideNavigationCondensedObjectExampleComponent } from './component-docs/side-navigation/examples/side-navigation-condensed-object-example/side-navigation-condensed-object-example.component';
 import { SideNavigationMultipleSelectedExampleComponent } from './component-docs/side-navigation/examples/side-navigation-multiple-selected-example/side-navigation-multiple-selected-example.component';
+import { ShellbarSideNavExampleComponent } from './component-docs/shellbar/examples/shellbar-side-nav/shellbar-side-nav-example.component';
 import { DatePickerComplexI18nExampleComponent } from './component-docs/date-picker/examples/date-picker-complex-i18n-example/date-picker-complex-i18n-example.component';
 import { DatetimePickerComplexI18nExampleComponent } from './component-docs/datetime-picker/examples/datetime-picker-complex-i18n-example/datetime-picker-complex-i18n-example.component';
+import { PopoverDropdownComponent } from './component-docs/popover/examples/popover-dropdown/popover-dropdown.component';
+import { SelectTypesExampleComponent } from './component-docs/select/examples/select-types-example/select-types-example.component';
 
 
 @NgModule({
@@ -594,6 +597,7 @@ import { DatetimePickerComplexI18nExampleComponent } from './component-docs/date
         ShellbarBasicExampleComponent,
         ShellbarAdvancedExampleComponent,
         ShellbarCollapsibleExampleComponent,
+        ShellbarSideNavExampleComponent,
         SideNavigationExampleComponent,
         SideNavigationIconsExampleComponent,
         SideNavigationTitlesExampleComponent,
@@ -733,6 +737,7 @@ import { DatetimePickerComplexI18nExampleComponent } from './component-docs/date
         SelectFormsComponent,
         SelectViewValueExampleComponent,
         SelectMaxHeightExampleComponent,
+        SelectTypesExampleComponent,
         CalendarI18nMomentExampleComponent,
         SelectMaxHeightExampleComponent,
         NotificationDocsComponent,
@@ -750,7 +755,8 @@ import { DatetimePickerComplexI18nExampleComponent } from './component-docs/date
         SideNavigationThreeLevelsExampleComponent,
         NotificationContentComponent,
         DatePickerComplexI18nExampleComponent,
-        DatetimePickerComplexI18nExampleComponent
+        DatetimePickerComplexI18nExampleComponent,
+        PopoverDropdownComponent
     ],
 
     entryComponents: [